+++ R/utils.R
@@ -21,7 +21,8 @@ globalVariables(
     "confirmed", "death", "desc", "recovered", "state", "cityName",
     "china_cities_location", "cities", "lat", "lon", "value",
     "confirmedCount", "curedCount", "deadCount", "suspectedCount",
-    "province", "province_pinyin", "variable", "chinese_provinces"
+    "province", "province_pinyin", "variable", "chinese_provinces",
+    "date2", "rate"
   )
 )