+++ dev/02_dev.R
@@ -19,6 +19,7 @@ golem::add_module( name = "city_map" ) # Name of the module
 golem::add_module( name = "dxy_table" ) # Name of the module
 golem::add_module( name = "geo_lines" ) # Name of the module
 golem::add_module( name = "time_provinces" ) # Name of the module
+golem::add_module( name = "jhu_death_rate" ) # Name of the module
 
 ## 2.2 Add dependencies