+++ docs/news.md
@@ -5,6 +5,7 @@
 - Extend toast duration to 5 seconds.
 - Remove coordinates of cities from crawl.
 - Switch to using [Github](https://github.com/CSSEGISandData/2019-nCoV) instead of spreadsheet for JHU data. 
+- Add death rate and plot to JHU dashboard
 
 # coronavirus 0.0.4